Revamp a little to make these tests work together.
